Finish button functionality and styling
Previously, the remove book button functionality made use of splice
which led to unforeseen outcomes.

Now, the app uses the findIndex array method to successfully
remove the right book from both the displayed card as well as
the array of book objects. Moreover, the app uses the find array
method to toggle the read status of a given book in the array
and then displays such status in the card, too.