Remove unnecessary data attributes
Previously, the app used data attributes to link displayed books
and the books in the book array.

Given that the app now uses unique book attributes (title, ...) to
find the right book to be removed or to have its read status
changed, the data attributes have been rendered obsolete and, thus,
removed.